Bug 2059762 - Don't reopen the destination file when copying mount content
We already get a file handle with `CreateFileAsTaskUser` which ensures
that the file we're copying to is one the task user has access to, so
avoid re-opening the file by file name which is prone to issues with
links being swapped at runtime by a second task or a stray process.
